CrateDB loader: SQLAlchemyLoader has been superseded by SQLDatabaseLoader
