+++ docs/docs/.gitignore
@@ -5,3 +5,4 @@ node_modules/
 .docusaurus
 .cache-loader
 docs/api
+example.sqlite
