+++ docs/docs/how_to/sql_database.mdx
@@ -1,10 +1,11 @@
-# SQLAlchemy
+# SQLDatabaseLoader
 
 
 ## About
 
-The [SQLAlchemy] document loader loads records from any supported database,
-see [SQLAlchemy dialects] for all supported SQL databases and dialects.
+The `SQLDatabaseLoader` loads records from any database supported by
+[SQLAlchemy], see [SQLAlchemy dialects] for the whole list of supported
+SQL databases and dialects.
 
 You can either use plain SQL for querying, or use an SQLAlchemy `Select`
 statement object, if you are using SQLAlchemy-Core or -ORM.
@@ -31,11 +32,11 @@ psql postgresql://postgres@localhost/testdrive < ./libs/langchain/tests/integrat
 ### Basic loading
 
 ```python
-from langchain.document_loaders import SQLAlchemyLoader
+from langchain_community.document_loaders.sql_database import SQLDatabaseLoader
 from pprint import pprint
 
 
-loader = SQLAlchemyLoader(
+loader = SQLDatabaseLoader(
     query="SELECT * FROM mlb_teams_2012 LIMIT 3;",
     url="postgresql+psycopg2://postgres@localhost:5432/testdrive",
 )
@@ -66,7 +67,7 @@ Having the `query` within metadata is useful when using documents loaded from
 database tables for chains that answer questions using their origin queries.
 
 ```python
-loader = SQLAlchemyLoader(
+loader = SQLDatabaseLoader(
     query="SELECT * FROM mlb_teams_2012 LIMIT 3;",
     url="postgresql+psycopg2://postgres@localhost:5432/testdrive",
     include_rownum_into_metadata=True,
@@ -97,11 +98,20 @@ the `metadata` dictionary with corresponding information. When `page_content_col
 is empty, all columns will be used.
 
 ```python
-loader = SQLAlchemyLoader(
+import functools
+
+row_to_content = functools.partial(
+    SQLDatabaseLoader.page_content_default_mapper, column_names=["Payroll (millions)", "Wins"]
+)
+row_to_metadata = functools.partial(
+    SQLDatabaseLoader.metadata_default_mapper, column_names=["Team"]
+)
+
+loader = SQLDatabaseLoader(
     query="SELECT * FROM mlb_teams_2012 LIMIT 3;",
     url="postgresql+psycopg2://postgres@localhost:5432/testdrive",
-    page_content_columns=["Payroll (millions)", "Wins"],
-    metadata_columns=["Team"],
+    page_content_mapper=row_to_content,
+    metadata_mapper=row_to_metadata,
 )
 docs = loader.load()
 ```
@@ -128,10 +138,10 @@ document created from each row. This is useful for identifying documents through
 their metadata. Typically, you may use the primary key column(s) for that purpose.
 
 ```python
-loader = SQLAlchemyLoader(
+loader = SQLDatabaseLoader(
     query="SELECT * FROM mlb_teams_2012 LIMIT 3;",
     url="postgresql+psycopg2://postgres@localhost:5432/testdrive",
-    source_columns="Team",
+    source_columns=["Team"],
 )
 docs = loader.load()
 ```
