+++ libs/community/tests/unit_tests/test_sql_database.py
@@ -3,6 +3,8 @@
 
 import pytest
 import sqlalchemy as sa
+import sqlalchemy.orm
+from langchain_community.utilities.sql_database import SQLDatabase, truncate_word
 from packaging import version
 from sqlalchemy import (
     Column,
@@ -15,8 +17,6 @@ from sqlalchemy import (
     select,
 )
 from sqlalchemy.engine import Engine, Result
-
-from langchain_community.utilities.sql_database import SQLDatabase, truncate_word
 
 is_sqlalchemy_v1 = version.parse(sa.__version__).major == 1
 
@@ -56,6 +56,11 @@ def db_lazy_reflection(engine: Engine) -> SQLDatabase:
 
 
 @pytest.mark.xfail(is_sqlalchemy_v1, reason="SQLAlchemy 1.x issues")
+def test_configure_mappers() -> None:
+    """Test that configuring table mappers works."""
+    sqlalchemy.orm.configure_mappers()
+
+
 def test_table_info(db: SQLDatabase) -> None:
     """Test that table info is constructed properly."""
     output = db.table_info
