--- libs/community/tests/unit_tests/test_sqlalchemy.py
@@ -1,7 +1,0 @@
-import sqlalchemy.orm
-
-import langchain_community  # noqa: F401
-
-
-def test_configure_mappers() -> None:
-    sqlalchemy.orm.configure_mappers()
