+++ libs/langchain/langchain/document_loaders/__init__.py
@@ -512,7 +512,6 @@ __all__ = [
     "SlackDirectoryLoader",
     "SnowflakeLoader",
     "SpreedlyLoader",
-    "SQLAlchemyLoader",
     "StripeLoader",
     "TelegramChatApiLoader",
     "TelegramChatFileLoader",
